refactor(racing-car): getter가 아닌 get으로 시작하는 메소드 이름 변경
